# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NotificationSettingsResponse": {
        "properties": {
          "daily_summary_enabled": {
            "title": "Daily Summary Enabled",
            "type": "boolean"
          },
          "diaper_reminder_enabled": {
            "title": "Diaper Reminder Enabled",
            "type": "boolean"
          },
          "dnd_end_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dnd End Time"
          },
          "dnd_start_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dnd Start Time"
          },
          "family_record_enabled": {
            "title": "Family Record Enabled",
            "type": "boolean"
          },
          "feeding_reminder_enabled": {
            "title": "Feeding Reminder Enabled",
            "type": "boolean"
          },
          "system_notice_enabled": {
            "title": "System Notice Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "family_record_enabled",
          "feeding_reminder_enabled",
          "diaper_reminder_enabled",
          "daily_summary_enabled",
          "system_notice_enabled"
        ],
        "title": "NotificationSettingsResponse",
        "type": "object"
      },
      "NotificationSettingsUpdate": {
        "properties": {
          "daily_summary_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Summary Enabled"
          },
          "diaper_reminder_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diaper Reminder Enabled"
          },
          "dnd_end_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dnd End Time"
          },
          "dnd_start_time": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dnd Start Time"
          },
          "family_record_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Record Enabled"
          },
          "feeding_reminder_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feeding Reminder Enabled"
          },
          "system_notice_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "System Notice Enabled"
          }
        },
        "title": "NotificationSettingsUpdate",
        "type": "object"
      },
...
      "PushSubscriptionCreate": {
        "properties": {
          "auth": {
            "title": "Auth",
            "type": "string"
          },
          "endpoint": {
            "title": "Endpoint",
            "type": "string"
          },
          "p256dh": {
            "title": "P256Dh",
            "type": "string"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          }
        },
        "required": [
          "endpoint",
          "p256dh",
          "auth"
        ],
        "title": "PushSubscriptionCreate",
        "type": "object"
      },
      "PushSubscriptionResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "endpoint": {
            "title": "Endpoint",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "endpoint",
          "created_at"
        ],
        "title": "PushSubscriptionResponse",
        "type": "object"
      },
...
      "ScheduleCreate": {
        "properties": {
          "baby_id": {
            "title": "Baby Id",
            "type": "integer"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_completed": {
            "default": false,
            "title": "Is Completed",
            "type": "boolean"
          },
          "scheduled_time": {
            "format": "date-time",
            "title": "Scheduled Time",
            "type": "string"
          },
          "title": {
            "maxLength": 100,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "baby_id",
          "title",
          "scheduled_time"
        ],
        "title": "ScheduleCreate",
        "type": "object"
      },
      "ScheduleResponse": {
        "properties": {
          "baby_id": {
            "title": "Baby Id",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_completed": {
            "default": false,
            "title": "Is Completed",
            "type": "boolean"
          },
          "scheduled_time": {
            "format": "date-time",
            "title": "Scheduled Time",
            "type": "string"
          },
          "title": {
            "maxLength": 100,
            "title": "Title",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "baby_id",
          "title",
          "scheduled_time",
          "id",
          "user_id",
          "created_at"
        ],
        "title": "ScheduleResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Botoro API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/notifications/settings": {
      "get": {
        "operationId": "get_settings_api_notifications_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Settings",
        "tags": [
          "notifications"
        ]
      },
      "patch": {
        "operationId": "update_settings_api_notifications_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Settings",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/notifications/subscribe": {
      "post": {
        "operationId": "subscribe_api_notifications_subscribe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscriptionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushSubscriptionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Subscribe",
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/api/notifications/unsubscribe": {
      "post": {
        "operationId": "unsubscribe_api_notifications_unsubscribe_post",
        "parameters": [
          {
            "in": "query",
            "name": "endpoint",
            "required": true,
            "schema": {
              "title": "Endpoint",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unsubscribe",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/notifications/{notification_id}/read": {
      "patch": {
        "description": "個別通知を既読にする",
        "operationId": "mark_as_read_api_notifications__notification_id__read_patch",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "title": "Notification Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark As Read",
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/api/schedules/": {
      "get": {
        "operationId": "get_schedules_api_schedules__get",
        "parameters": [
          {
            "in": "query",
            "name": "baby_id",
            "required": true,
            "schema": {
              "title": "Baby Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScheduleResponse"
                  },
                  "title": "Response Get Schedules Api Schedules  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Schedules",
        "tags": [
          "schedules"
        ]
      },
      "post": {
        "operationId": "create_schedule_api_schedules__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Schedule",
        "tags": [
          "schedules"
        ]
      }
    },
    "/api/schedules/{schedule_id}": {
      "delete": {
        "operationId": "delete_schedule_api_schedules__schedule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Schedule",
        "tags": [
          "schedules"
        ]
      }
    },
...
  }
}